+++ dmeta/errors.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""DMeta errors."""
+
+
+class DMetaBaseError(Exception):
+    """DMeta base error class."""
+
+    pass