+++ dmeta/util.py
@@ -5,7 +5,9 @@ import json
 from shutil import rmtree
 from zipfile import ZipFile
 import defusedxml.ElementTree as ET
-from .params import SUPPORTED_MICROSOFT_FORMATS
+from .params import SUPPORTED_MICROSOFT_FORMATS, NOT_IMPLEMENTED_ERROR, \
+    FILE_FORMAT_DOES_NOT_EXIST_ERROR, INVALID_CONFIG_FILE_NAME_ERROR, CONFIG_FILE_DOES_NOT_EXIST_ERROR
+from .errors import DMetaBaseError
 
 
 def extract_namespaces(xml_file_path):
@@ -39,10 +41,10 @@ def get_microsoft_format(file_name):
     """
     last_dot_index = file_name.rfind('.')
     if (last_dot_index == -1):
-        return None
+        raise DMetaBaseError(FILE_FORMAT_DOES_NOT_EXIST_ERROR)
     format = file_name[last_dot_index + 1:]
     if format not in SUPPORTED_MICROSOFT_FORMATS:
-        return None
+        raise DMetaBaseError(NOT_IMPLEMENTED_ERROR)
     return format
 
 
@@ -71,11 +73,11 @@ def read_json(config_file_name):
     :return: obj
     """
     if not isinstance(config_file_name, str):
-        raise ("Given config file name should be str not the other type.")
+        raise DMetaBaseError(INVALID_CONFIG_FILE_NAME_ERROR)
     if ".json" not in config_file_name:
         config_file_name = config_file_name + ".json"
     if os.path.isfile(config_file_name):
         config_file = open(config_file_name)
         return json.load(config_file)
     else:
-        raise ("Given config file doesn't exist.")
+        raise DMetaBaseError(CONFIG_FILE_DOES_NOT_EXIST_ERROR)